pyproject.toml
[project]
name = "TTBoard"
version = "0.0.1"
description = "GUI application to edit database and JSON files"
readme = "README.md"

dependencies = [
  "typer",
  "python-jsonpath",
  "pillow",
  "pytest"
  ]

[project.scripts]
ttboard-cli = "ttboard.cli.main:main"
ttboard-test-jsonpath = "ttboard.tests.testJsonPath:main"
ttboard-test-selector = "ttboard.tests.testSelector:main"
ttboard-test-appflow1 = "ttboard.tests.testAppFlow1:main"
ttboard-test-update = "ttboard.tests.testUpdate:main"

[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [ "src/ttboard" ]
include = ["src/ttboard/assets" ]

[tool.pytest.ini_options]
testpaths = [ "tests" ]
pythonpath = [ "src" ]
filterwarnings = [ "ignore::DeprecationWarning" ]
